converts Selector booleans to strings
`True` becomes `'t'` and `False` becomes `'f'`. In the future we can
make this mapping overridable if necessary, but for now we know we
need this (and it's not hard to see how you'd replace that `dict` I'm
using with an environmentally determined constant), so we'll keep it
simple for now.